refactoring twitter tools structure
