+++ src/agents/workflows/twitter/tools.ts
@@ -1,4 +1,4 @@
-import { createAllTwitterTools } from '../../tools/twitter/twitterTools.js';
+import { createAllTwitterTools } from '../../tools/twitter/index.js';
 import { TwitterApi } from '../../../services/twitter/types.js';
 import { createSaveExperienceTool } from '../../tools/dsn/saveExperienceTool.js';
 import { createGetCurrentTimeTool } from '../../tools/time/getTimeTool.js';
